requirements.txt
weasyprint
pymupdf
Pillow
